requirements.txt
flask
requests
python-dotenv
geopy
json
jsonify
sys
ssl
certifi

